SpeechTranscriberをドメイン層から削除 #48
